+++ graph/node_pool_test.go
@@ -69,7 +69,7 @@ func TestMaxPool(t *testing.T) {
 }
 
 func TestGradientMaxPool(t *testing.T) {
-	testGradients[float32](t, "Gradient 1D MaxPool().NoPadding() -- scaled output",
+	testGradients(t, "Gradient 1D MaxPool().NoPadding() -- scaled output",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			input := Add(IotaFull(g, MakeShape(shapes.Float32, 1, 6, 1)), Const(g, float32(1.0)))
 			output = MaxPool(input).NoPadding().Window(3).Strides(1).Done()
@@ -80,7 +80,7 @@ func TestGradientMaxPool(t *testing.T) {
 			[][][]float32{{{0}, {0}, {1}, {2}, {3}, {4}}},
 		})
 
-	testGradients[float64](t, "Gradient 1D MaxPool(...).Window(2).Strides(1)",
+	testGradients(t, "Gradient 1D MaxPool(...).Window(2).Strides(1)",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			channelA := IotaFull(g, MakeShape(shapes.Float64, 1, 5, 1))
 			channelB := Mul(channelA, Const(g, 0.1))
@@ -91,7 +91,7 @@ func TestGradientMaxPool(t *testing.T) {
 			[][][]float64{{{0, 0}, {1, 1}, {1, 1}, {1, 1}, {1, 1}}},
 		})
 
-	testGradients[float64](t, "Gradient 2D MaxPool(...).Window(2)",
+	testGradients(t, "Gradient 2D MaxPool(...).Window(2)",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			input := IotaFull(g, MakeShape(shapes.Float64, 1, 5, 5, 1))
 			output = MaxPool(input).Window(2).Done()
@@ -123,7 +123,7 @@ func TestSumPool(t *testing.T) {
 // TestGradientSumPool only tests that the correct gradient for the reduction is applied. More fine-grained testing
 // is done in TestGradientMeanPool, which we can verify the result with Tensorflow.
 func TestGradientSumPool(t *testing.T) {
-	testGradients[float32](t, "Gradient 2D SumPool(...).Window(3).PadSame().Strides(1)",
+	testGradients(t, "Gradient 2D SumPool(...).Window(3).PadSame().Strides(1)",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			input := IotaFull(g, MakeShape(shapes.Float32, 1, 3, 1))
 			output = SumPool(input).Window(3).PadSame().Strides(1).Done()
@@ -132,7 +132,7 @@ func TestGradientSumPool(t *testing.T) {
 			[][][]float32{{{2}, {3}, {2}}},
 		})
 
-	testGradients[float64](t, "Gradient 2D Even Spatial Dimensions SumPool(...).Window(2)",
+	testGradients(t, "Gradient 2D Even Spatial Dimensions SumPool(...).Window(2)",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			input := IotaFull(g, MakeShape(shapes.Float64, 1, 5, 1))
 			output = SumPool(input).Window(2).Strides(3).
@@ -143,7 +143,7 @@ func TestGradientSumPool(t *testing.T) {
 			[][][]float64{{{1}, {1}, {0}, {2}, {2}}},
 		})
 
-	testGradients[float64](t, "Gradient 2D Even Spatial Dimensions SumPool(...).Window(2)",
+	testGradients(t, "Gradient 2D Even Spatial Dimensions SumPool(...).Window(2)",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			input := IotaFull(g, MakeShape(shapes.Float64, 1, 4, 4, 1))
 			output = SumPool(input).Window(2).Done()
@@ -158,7 +158,7 @@ func TestGradientSumPool(t *testing.T) {
 			}},
 		})
 
-	testGradients[float32](t, "Gradient 1D SumPool() -- scaled output",
+	testGradients(t, "Gradient 1D SumPool() -- scaled output",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			input := Add(IotaFull(g, MakeShape(shapes.Float32, 1, 6, 1)), Const(g, float32(1.0)))
 			output = SumPool(input).NoPadding().Window(3).Strides(1).Done()
@@ -205,7 +205,7 @@ func TestMeanPool(t *testing.T) {
 // TestGradientMeanPool can leverage `tensorflow.nn.pool` with "AVG" type. Results should be identical.
 func TestGradientMeanPool(t *testing.T) {
 	// The results below should reflect that the last column and row are not used in the output.
-	testGradients[float64](t, "Gradient 2D MeanPool(...).Window(2)",
+	testGradients(t, "Gradient 2D MeanPool(...).Window(2)",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			input := IotaFull(g, MakeShape(shapes.Float64, 1, 5, 5, 1))
 			output = MeanPool(input).Window(2).Done()
@@ -220,7 +220,7 @@ func TestGradientMeanPool(t *testing.T) {
 			}},
 		})
 
-	testGradients[float64](t, "Gradient Unbalanced 1D MeanPool(...).Window(2)",
+	testGradients(t, "Gradient Unbalanced 1D MeanPool(...).Window(2)",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			input := IotaFull(g, MakeShape(shapes.Float64, 1, 3, 1))
 			output = MeanPool(input).Window(2).PaddingPerDim([][2]int{{2, 0}}).Done()
@@ -229,7 +229,7 @@ func TestGradientMeanPool(t *testing.T) {
 			[][][]float64{{{0.5}, {0.5}, {0.0}}},
 		})
 
-	testGradients[float32](t, "Gradient 1D MeanPool() -- scaled output",
+	testGradients(t, "Gradient 1D MeanPool() -- scaled output",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			input := Add(IotaFull(g, MakeShape(shapes.Float32, 1, 6, 1)), Const(g, float32(1.0)))
 			output = MeanPool(input).NoPadding().Window(3).Strides(1).Done()
@@ -241,7 +241,7 @@ func TestGradientMeanPool(t *testing.T) {
 		})
 	fmt.Println()
 
-	testGradients[float32](t, "Gradient 2D MeanPool(...).Window(3).ChannelsFirst().PadSame().Strides(1)",
+	testGradients(t, "Gradient 2D MeanPool(...).Window(3).ChannelsFirst().PadSame().Strides(1)",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			channelA := IotaFull(g, MakeShape(shapes.Float32, 1, 1, 3, 3))
 			channelB := Mul(channelA, Const(g, float32(0.1)))
@@ -261,7 +261,7 @@ func TestGradientMeanPool(t *testing.T) {
 		})
 	fmt.Println()
 
-	testGradients[float64](t, "Gradient 2D even spatial dims MeanPool(...).Window(2)",
+	testGradients(t, "Gradient 2D even spatial dims MeanPool(...).Window(2)",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			input := IotaFull(g, MakeShape(shapes.Float64, 1, 2, 2, 1))
 			output = MeanPool(input).Window(2).Done()