+++ docs/CHANGELOG.md
@@ -1,5 +1,16 @@
 # GoMLX changelog
 
+## v0.6.0 - 2023/08/07
+
+* FFT, RealFFT, InverseFFT and InverseRealFFT operations.
+  * Added a small notebook demo for FFT. 
+* Added Complex/Imag/Real/Conj operations to manipulate complex numbers (and their gradients).
+* Added support for complex numbers for ConvertType. Defined gradient for ConvertType.
+* Added Complex128 and Complex64 dtypes support.
+* Added "spacers" (like "*" for axis ranges) and `AxisElem()` for `Slice()`.
+* Package `examples/notebook/gonb/margaid`: Added `Plots.AddValues` and `Plots.PlotToHTML`; 
+  Fixed `Plots` returned by `New()` to be linear scale by default.
+
 ## v0.5.0 - 2023/07/10
 
 * Error handling revamp: using `panic` to report errors — it works as exceptions. This is a very large change